Move everything to modal folder
