+++ src/components/info/reminder.tsx
@@ -1,7 +1,7 @@
 import { CardHeader } from 'components/info/card'
 import { NoteDisplay, NoteInput, NoteMenu } from 'components/info/note'
 import { VisibilityToggle } from 'components/info/visibilityToggle'
-import DeleteConfirmModal from 'components/input/generic_modals/delete_confirm_modal'
+import DeleteConfirmModal from 'components/modals/delete_confirm_modal'
 import { useAppStatus } from 'context/useAppStatus'
 import { useSavedArmies } from 'context/useSavedArmies'
 import { useSubscription } from 'context/useSubscription'
