+++ src/components/input/savedArmies/save_army_btn.tsx
@@ -1,8 +1,8 @@
 import { useAuth0 } from '@auth0/auth0-react'
 import { OfflineBtn } from 'components/helpers/suspenseFallbacks'
 import GenericButton from 'components/input/generic_button'
-import SubscriberOnlyModal from 'components/input/generic_modals/subscriber_only_modal'
-import { SaveArmyModal } from 'components/input/savedArmies/save_army_modal'
+import { SaveArmyModal } from 'components/modals/save_army_modal'
+import SubscriberOnlyModal from 'components/modals/subscriber_only_modal'
 import { useAppStatus } from 'context/useAppStatus'
 import { useSubscription } from 'context/useSubscription'
 import React, { useState } from 'react'
