+++ src/components/input/savedArmies/saved_army_card.tsx
@@ -1,8 +1,8 @@
 import GenericButton from 'components/input/generic_button'
-import { DeleteArmyModal } from 'components/input/savedArmies/delete_army_modal'
 import { LoadArmyBtn } from 'components/input/savedArmies/load_army_btn'
 import { SavedArmyTable } from 'components/input/savedArmies/saved_army_table'
 import UpdateNameButton from 'components/input/savedArmies/update_name_btn'
+import { DeleteArmyModal } from 'components/modals/delete_army_modal'
 import { useAppStatus } from 'context/useAppStatus'
 import { useSavedArmies } from 'context/useSavedArmies'
 import { useTheme } from 'context/useTheme'
