+++ src/components/input/shareArmy/share_army_btn.tsx
@@ -1,7 +1,7 @@
 import { useAuth0 } from '@auth0/auth0-react'
 import GenericButton from 'components/input/generic_button'
-import SubscriberOnlyModal from 'components/input/generic_modals/subscriber_only_modal'
-import { ShareArmyModal } from 'components/input/shareArmy/share_army_modal'
+import { ShareArmyModal } from 'components/modals/share_army_modal'
+import SubscriberOnlyModal from 'components/modals/subscriber_only_modal'
 import { useAppStatus } from 'context/useAppStatus'
 import { useSubscription } from 'context/useSubscription'
 import React, { useState } from 'react'
