+++ src/components/modals/delete_confirm_modal.tsx
@@ -1,5 +1,5 @@
 import GenericButton from 'components/input/generic_button'
-import GenericModal from 'components/input/generic_modals/genericModal'
+import GenericModal from 'components/modals/generic_modal'
 import { useTheme } from 'context/useTheme'
 import React from 'react'
 import { FaCheck } from 'react-icons/fa'
