+++ src/components/modals/generic_modal.tsx
@@ -27,7 +27,7 @@ const GenericModal: React.FC<IGenericModalProps> = props => {
       onRequestClose={closeModal}
       overlayClassName={'Modal-Overlay'}
     >
-      <div className={`container ${isProcessing ? `` : `mr-3 pl-0`}`}>
+      <div className={`container`}>
         {isProcessing && <ModalSpinner />}
         <div hidden={isProcessing}>{children}</div>
       </div>
