+++ src/components/modals/share_army_modal.tsx
@@ -1,6 +1,6 @@
 import Spinner from 'components/helpers/spinner'
 import GenericButton from 'components/input/generic_button'
-import GenericModal from 'components/input/generic_modals/genericModal'
+import GenericModal from 'components/modals/generic_modal'
 import { useSavedArmies } from 'context/useSavedArmies'
 import { useTheme } from 'context/useTheme'
 import { selectors } from 'ducks'
