+++ src/components/modals/stripe_cancellation_modal.tsx
@@ -1,32 +1,29 @@
-import { LinkNewTab } from 'components/helpers/link'
 import GenericButton from 'components/input/generic_button'
-import GenericModal from 'components/input/generic_modals/genericModal'
+import GenericModal from 'components/modals/generic_modal'
 import { useSubscription } from 'context/useSubscription'
 import { useTheme } from 'context/useTheme'
 import React, { useState } from 'react'
 import { IconContext } from 'react-icons'
 import { FaRegSadCry } from 'react-icons/fa'
 import { logClick } from 'utils/analytics'
-import { isDev } from 'utils/env'
 
 interface IModalComponentProps {
   modalIsOpen: boolean
   closeModal: () => void
 }
 
-export const CancelPaypalSubscriptionModal: React.FC<IModalComponentProps> = props => {
+export const CancelStripeSubscriptionModal: React.FC<IModalComponentProps> = props => {
   const { closeModal, modalIsOpen } = props
-  const { subscription } = useSubscription()
+  const { cancelSubscription } = useSubscription()
   const { theme } = useTheme()
   const [processing, setProcessing] = useState(false)
 
-  const handleClick = () => {
+  const handleClick = async (e: React.MouseEvent) => {
+    e.preventDefault()
     setProcessing(true)
+    await cancelSubscription()
     logClick('CancelSubscription')
-    setTimeout(() => {
-      setProcessing(false)
-      closeModal()
-    }, 3000)
+    setProcessing(false)
   }
 
   return (
@@ -46,28 +43,15 @@ export const CancelPaypalSubscriptionModal: React.FC<IModalComponentProps> = pro
             <p>
               You'll still have access to all subscription features until your current subscription expires.
             </p>
-            <p>
-              <small>
-                You will be redirected to cancel your subscription using the Paypal interface for extra
-                security. The change in your subscription may not be reflected for up to five minutes.
-              </small>
-            </p>
           </div>
         </IconContext.Provider>
       </div>
 
       <div className="row text-center">
         <div className="col px-0">
-          <LinkNewTab
-            href={`https://www.${isDev ? 'sandbox.' : ''}paypal.com/myaccount/autopay/connect/${
-              subscription.subscriptionId
-            }`}
-            label={'Paypal Unsubscribe'}
-          >
-            <GenericButton className={theme.modalDangerClass} onClick={handleClick}>
-              Cancel Subscription
-            </GenericButton>
-          </LinkNewTab>
+          <GenericButton className={theme.modalDangerClass} onClick={handleClick}>
+            Cancel Subscription
+          </GenericButton>
 
           <GenericButton className={theme.modalConfirmClass} onClick={closeModal}>
             Back
