+++ src/components/payment/pricingPlans.tsx
@@ -3,7 +3,7 @@ import { Elements, useStripe } from '@stripe/react-stripe-js'
 import { loadStripe } from '@stripe/stripe-js'
 import { SubscriptionApi } from 'api/subscriptionApi'
 import GenericButton from 'components/input/generic_button'
-import { PaypalPostSubscribeModal } from 'components/input/paypal_post_subscribe_modal'
+import { PaypalPostSubscribeModal } from 'components/modals/paypal_post_subscribe_modal'
 import PayPalButton from 'components/payment/paypal/paypalButton'
 import { PaypalProvider } from 'context/usePaypal'
 import qs from 'qs'
