+++ src/components/routes/Profile.tsx
@@ -1,9 +1,9 @@
 import { useAuth0 } from '@auth0/auth0-react'
 import { LoadingBody, LoadingHeader } from 'components/helpers/suspenseFallbacks'
 import GenericButton from 'components/input/generic_button'
-import { CancelPaypalSubscriptionModal } from 'components/input/paypal_cancellation_modal'
 import { SelectOne } from 'components/input/select'
-import { CancelStripeSubscriptionModal } from 'components/input/stripe_cancellation_modal'
+import { CancelPaypalSubscriptionModal } from 'components/modals/paypal_cancellation_modal'
+import { CancelStripeSubscriptionModal } from 'components/modals/stripe_cancellation_modal'
 import Contact from 'components/page/contact'
 import { GiftSubscriptions } from 'components/payment/giftSubscriptions'
 import { useSavedArmies } from 'context/useSavedArmies'
